Add `--example` flag to `crane build` and `crane run` (#26)
This PR adds an `--example` flag to `crane build` and `crane run` to
allow specifying which program in the `examples/` directory should be
built or run.

This helps remove some of the hard-coding of `scratch.crane` everywhere
(although that is still the default program that gets run if `--example`
is not provided.

```
$ crane build --help
Compiles the current project

Usage: crane build [OPTIONS]

Options:
      --example <EXAMPLE>  Builds the given example
  -h, --help               Print help
```

```
$ crane run --help
Runs the current project

Usage: crane run [OPTIONS]

Options:
      --example <EXAMPLE>  Runs the given example
  -h, --help               Print help
```