Use paths in typed AST (#33)
This PR updates the typed AST to use `TyPath`s instead of `Ident`s for
variables.

This allows us to refer to functions defined in other (inline) modules.

However, it does come at the cost of having to prefix all of the
standard library function calls with the appropriate path prefix. This
is just a temporary thing, as once `use` items are implemented it will
be possible to bring these functions into scope.

The code powering this is... rough, to put it nicely. Building up and
keeping track of absolute paths to items is rather hard to do in an
ad-hoc way. I suspect we'll need to redo this very soon.