Distinguish between comments and doc comments during lexing (#39)
This PR makes the lexer treat comments (`//`) and doc comments (`///`)
as different tokens.

At present this doesn't make any difference from a parsing perspective,
as doc comments were already parsable, but just as regular comments.

In the future, however, we will want to be able to distinguish between
the two so we can extract doc comments (e.g., in generated
documentation).