Store Inkwell objects on `NativeBackend` (#51)
This PR refactors the `NativeBackend` to store the various Inkwell
objects on it rather than threading them around as parameters
everywhere.

One small downside to this is that the caller needs to own the `Context`
and pass it in to the `NativeBackend` constructor, but I couldn't find a
could way to get around this without creating a self-referential struct.
While it would be nice to come up with a solve for this, it's pretty
minor at this point.